Set up jupytext
Version controlling jupyter notebooks is tedious, since cell
output is stored in base64 encoding.
Jupytext is a jupyter/jupyterlab-extension that pairs notebooks and
script files, such that changes can be reviewed easily. All .ipynb-files
and .py-files can be changed, the other one will be updated accordingly
if jupyter is running.

Look for the [Jupytext Server Extension]-line when starting the jupyter
server - if it is not there, follow the instructions in the README.