Add peak routing example
in this example the routes to different mountain peaks,
extracted from OSM with overpy, are calculated for three
starting points.
The resulting maps are transformed to PNGs using
selenium and webdriver-manager.
Optional image editing is done by example of a resize
using Pillow.

This example was contributed by Dr. Stephan Fuchs

Co-authored-by: Stephan <nicefuchs@gmail.com>